requirements.txt
amqp==2.5.2
beautifulsoup4==4.8.1
billiard==3.5.0.5
celery==4.1.0
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==2.0
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==1.1.1
django-hitcount==1.3.1
django-tracking2==0.5.0
idna==2.8
importlib-metadata==1.3.0
kombu==4.6.7
more-itertools==8.0.2
oauthlib==3.1.0
Pillow==6.2.1
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
simplejson==3.17.0
six==1.13.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sorl-thumbnail==12.5.0
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.25.7
vine==1.3.0
zipp==0.6.0
